backend/requirements.txt
# Framework et serveur
fastapi>=0.115.10
uvicorn>=0.34.0
starlette>=0.46.0
pydantic>=2.10.6
pydantic[email]>=2.10.6
pydantic-settings>=2.8.1
python-multipart>=0.0.20

# Base de données
mysql-connector-python>=9.2.0
SQLAlchemy>=2.0.38
alembic>=1.14.1

# Authentification et sécurité
PyJWT>=2.10.1
bcrypt>=4.3.0
passlib>=1.7.4
python-jose>=3.4.0

# Opérations HTTP et API externes
httpx>=0.28.1
aiohttp>=3.11.13
requests>=2.32.3

# Manipulation de données
pandas>=2.2.3
numpy>=2.2.3
scikit-learn>=1.6.1
python-dateutil>=2.9.0.post0

# Environnement et configuration
python-dotenv>=1.0.1
environs>=14.1.1

# Utilitaires
tqdm>=4.67.1
Jinja2>=3.1.5
PyYAML>=6.0.2

# Tests
pytest>=8.3.4
pytest-asyncio>=0.25.3
pytest-cov>=6.0.0